+++ src/main/java/com/lk/chat/SocketApplication.java
@@ -1,0 +1,13 @@
+package com.lk.chat;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class SocketApplication {
+
+    public static void main(String[] args) {
+        SpringApplication.run(SocketApplication.class);
+    }
+
+}
