+++ src/main/java/com/lk/chat/chatroom/repository/ChatRoomRepository.java
@@ -1,0 +1,11 @@
+package com.lk.chat.chatroom.repository;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+
+import java.util.List;
+
+public interface ChatRoomRepository extends JpaRepository<ChatRoom, Long> {
+
+    List<ChatRoom> findAllByRoomName(String roomName);
+
+}
