+++ src/main/java/com/lk/chat/chatroom/service/ChatRoomService.java
@@ -1,0 +1,34 @@
+package com.lk.chat.chatroom.service;
+
+import com.lk.chat.chatroom.dto.ChatRoomDto;
+import com.lk.chat.chatroom.repository.ChatRoomRepository;
+import lombok.RequiredArgsConstructor;
+import org.springframework.stereotype.Service;
+
+import java.util.List;
+import java.util.stream.Collectors;
+
+@RequiredArgsConstructor
+@Service
+public class ChatRoomService {
+
+    private final ChatRoomRepository chatRoomRepository;
+
+
+//    public ChatRoomDto findRoomById(String id){
+//        return chatRoomDTOMap.get(id);
+//    }
+
+    public List<ChatRoomDto> findAll() {
+        return chatRoomRepository.findAll().stream()
+                .map(ChatRoomDto::new)
+                .collect(Collectors.toList());
+    }
+
+    public List<ChatRoomDto> findAllByRoomName(String roomName) {
+        return chatRoomRepository.findAllByRoomName(roomName).stream()
+                .map(ChatRoomDto::new)
+                .collect(Collectors.toList());
+    }
+
+}
