+++ src/main/java/com/lk/chat/chatting/enums/MessageType.java
@@ -1,0 +1,13 @@
+package com.lk.chat.chatting.enums;
+
+import lombok.Getter;
+import lombok.RequiredArgsConstructor;
+
+@Getter
+@RequiredArgsConstructor
+public enum MessageType {
+    ERROR,
+    WARNING,
+    INFO,
+    NONE,
+}
