+++ src/main/java/com/lk/chat/config/H2ServerConfig.java
@@ -1,0 +1,24 @@
+package com.lk.chat.config;
+
+import com.zaxxer.hikari.HikariDataSource;
+import org.h2.tools.Server;
+import org.springframework.boot.context.properties.ConfigurationProperties;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.context.annotation.Profile;
+
+import javax.sql.DataSource;
+import java.sql.SQLException;
+
+@Configuration
+@Profile("test")
+public class H2ServerConfig {
+
+    @Bean
+    @ConfigurationProperties("spring.datasource.hikari")
+    public DataSource dataSource() throws SQLException {
+        Server.createTcpServer("-tcp", "-tcpAllowOthers", "-tcpPort", "8081").start();
+
+        return new HikariDataSource();
+    }
+}
