--- src/main/java/com/lk/socket/SocketApplication.java
@@ -1,13 +1,0 @@
-package com.lk.socket;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class SocketApplication {
-
-    public static void main(String[] args) {
-        SpringApplication.run(SocketApplication.class);
-    }
-
-}
