--- src/main/java/com/lk/socket/chatroom/dto/ChatRoomDto.java
@@ -1,42 +1,0 @@
-package com.lk.socket.chatroom.dto;
-
-import com.lk.socket.chatroom.repository.ChatRoom;
-import io.swagger.v3.oas.annotations.media.Schema;
-import lombok.AccessLevel;
-import lombok.Builder;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import org.springframework.web.socket.WebSocketSession;
-
-import java.util.HashSet;
-import java.util.Set;
-import java.util.UUID;
-
-@Schema(name = "chatRoomDto", description = "채팅방 Dto")
-@Getter
-@NoArgsConstructor(access = AccessLevel.PROTECTED)
-public class ChatRoomDto {
-    private String roomId;
-    private String name;
-    private Set<WebSocketSession> sessions;
-
-    @Builder
-    public ChatRoomDto(String roomId, String name) {
-        this.roomId = roomId;
-        this.name = name;
-        sessions = new HashSet<>();
-    }
-
-    public static ChatRoomDto create(String name) {
-        ChatRoomDto room = new ChatRoomDto();
-        room.roomId = UUID.randomUUID().toString();
-        room.name = name;
-        return room;
-    }
-
-    public ChatRoomDto(ChatRoom chatRoom) {
-        this.roomId = chatRoom.getRoomId();
-        this.name = chatRoom.getRoomName();
-    }
-
-}
