--- src/main/java/com/lk/socket/chatroom/repository/ChatRoomRepository.java
@@ -1,9 +1,0 @@
-package com.lk.socket.chatroom.repository;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface ChatRoomRepository extends JpaRepository<ChatRoom, Long> {
-
-
-
-}
