--- src/main/java/com/lk/socket/chatting/enums/MessageType.java
@@ -1,13 +1,0 @@
-package com.lk.socket.chatting.enums;
-
-import lombok.Getter;
-import lombok.RequiredArgsConstructor;
-
-@Getter
-@RequiredArgsConstructor
-public enum MessageType {
-    ERROR,
-    WARNING,
-    INFO,
-    NONE,
-}
