+++ src/main/resources/application-test.yml
@@ -1,9 +1,12 @@
 spring:
   datasource:
-    url: jdbc:h2:mem:testdb;MODE=MySQL;DB_CLOSE_DELAY=-1;DB_CLOSE_ON_EXIT=FALSE
-    username: sa
-    password:
-    driver-class-name: org.h2.Driver
+    hikari:
+      driver-class-name: org.h2.Driver
+      jdbc-url: jdbc:h2:tcp://backend.socket.com:8081/../chatting-db
+#      jdbc-url: jdbc:h2:../chatting-db
+#    url: jdbc:h2:../chatting-db;MODE=MySQL;DB_CLOSE_DELAY=-1;DB_CLOSE_ON_EXIT=FALSE
+      username: sa
+      password:
   h2.console.enabled: true
   jpa:
     hibernate:
